Refactor all import code into Imports module
Since there is so much commanality between all of the modules that back
the DB records, extract that code into the OOTPUtility.Imports module
and import it with the `use` macro.  Now, the contract to use the
OOTPUtility.Imports module is to:

1. `use OOTPUtility.Imports`
2. Call the macro `attributes_to_import` imported from the `use` call,
   to specify the columns that will be imported to the database.
3. If there are any transformation needed to the raw data coming from
the CSV file, implement the `sanitize_csv_data` method.
4. If there are any transformations that need to be made the data after
its extracted from the CSV file, implement the `sanitize_attributes`
method, which is mostly used for generating the correct composite IDs
for the records associated with this module.