Remove fk_key_constraints from imported tables
Since we are going to be importing a ton of data that already exists,
there is very little reaosn to include a fk_key_constraint on the fk_key
columns, so remove them all.  Now, the imported data can be brought in
out of any particular order.  Before this change, the data had to be
imported specific order to not violate the fk_key_constraints that
existed.

While I'm here, remove the teams reference to `city_id`, we won't be
using it for anything useful.