Use an Agent to store imported IDs during import
When we are importing all of the records from a league, there are
various records along the way that have foreign key constraints, which
up until now were verified before import by attempting a DB lookup of
the record in question, which results in 1000's of extra DB calls.

Instead, as we are importing a league, save the imported IDs for the
tables we are imoprting so that the other tables to be imported for the
league can look up their foreign keys in memory instead of making a DB
call each and every time.