All Standings components use Surface.LiveComponent
Since we will need to use the socket to link back and forth between
standings views, and even possibly link on to the team, update all of
the component to use `Surface.LiveComponent`.  This involves generating
an ID for the `OOTPUtility.Standings` modules and updating the render
code to specifiy an ID for the components at render time.