Update the way Players.Player handles position
Before this commit, we were storing two fields to effectively calculate
a player's "position", both the position column and the "role" column if
the position is "1".  Instead of storing both of these columns as coded
integers, just store the string abbreviation for the position in the
"position" column and remove the "role column".