Fix statistics issues, including correct rounding
Update the rounding of most of the statistics that are stored as floats,
to the accepted level of precision that they are displayed historically.

Also, sort out the naming and inclusion of both WAR and WPA, in the
player statistics so they match what we are importing and are named
consistently amongst all of the types of statistics.