Add Teams.get_roster/2 to get specific team roster
Add a method that will return a team's roster, the active one by
default, or another of your choosing, depending on the value of `type`
passed in.

As a side effect of the testing, create TeamRoster factories for
:active, :expanded and :injured rosters.

While those factories probably seem a little excessive, I am
intentionally try to be pretty rigourous in creating factories that
produce realistic data, since it will make the comoponent tests I am
planning on writing easier to understand/work with.