Fix ExMachina factories with nested resources
When we have nested resources that are used in the creation of children,
ExMachina seems to struggle propagating the IDs of the resources if we
assemble them using `build/1`.  So in the instances where we are passing
created resources down into a further factory, call `insert/1` instead
of `build/1` to faciliate the correct generation of the IDs of the
parent resources.