Refactor Statistics.Batting.for_player/2
Improve the way that query for player batting statistics, giving more
options and making the development of the query recursive. Updates test
to ensure they still pass after the refactor.

Also, update the players_batting_stats_factory to ensure that unless
specified, the relationships between a player, his team, its league and
its level are presevered so that the rule of least surprise applies,
unless they are specifically overriden in the call to `build/1` or
`insert/1`.