Correct scoping of query for Leaderboards
Refactor Statistics.for_team/2 to delegate to
Statistics.Batting.for_team/2 when asking for a batting statistic.  Then
update Statistics.Batting.for_team/2 to use the query for fetch a
player's current batting stats, along with the player, to ensure we are
returning the correct batting stats for each of the leader categories.