Add Snapshy to ComponentCase for snapshot testing
Add Snapshy to ComponentCase will allow use to easily setup some form of
snapshot testing with our components to make sure that they aren't
breaking without our knowledge.

I think for the most part this is enough testing for the components for
the time being.  All are relatively simple and once I am happy with the
design/html that shouldn't change all that much and its as simple as
overriding the snapshot when they do.

But, as I try to create more shared components for things like headers
and different typograpy and refactor the application code I want a quick
way to make sure I didn't break anything.

I'll most likely add more robust tests for the LiveViews that handle
events but for the components that make up those LiveViews this seems
like a decent failsafe approach.