Factory don't need to preload just created records
Probably as a indication I was still thinking about the structures as
objects instead of just structures, after calling a factory methods like
ConferenceFactory.with_conferences/2 I would preload the conferences I
just created into the struct I have a reference to.

Instead of doing this, just update the struct we have the reference to
with the key of the records with just created and assign those records
to that key.