Refactor with_conferences/2 and with_divisions/2
When dealing with snapshot tests, its important that the generated data
being input into the component is consistent, at least for the fields
that are used by the component being tested.  This presented an issue
with teh current implemenation of `with_conferences/2` and
`with_divisions/2`, since there was no way to set individual
characteristics of the conferences or divisions being created, only
attributes that were shared amongst the children being created (league
for example) could be passed in.

This update changes the API to expect an array of attribute maps as the
second argument to the function, defaulting to [%{}, %{}] to match the
behavior of calling the existing `with_conferences` or `with_divisions`
functions.

However, that behavior now allows us to specifiy specific attributes for
each of the children being created (like name, for example) to allow us
to build data structures that remain constant from test rendering to
test rendering, but maintains the ability to create multiple versions of
the child without trouble if nothign more is required.