Extract Team.NameWithLogo component
Extract a component that represent the team name with its logo, which
can be used in a number of different scenarios, but now mostly
importantly in the PlayerLive view.

Fix a bug in the Facotry code that was incorrectly generating slugs for
Players and Teams, by only slugifying a "name" field, so add matches for
teams (name, nickname) and player (first_name, last_name).