Fix method signature of Teams.get_organization/2
The spec for the signature of Teams.get_organization/2 was incorrect as
it was being used, which was originally assumed just to be a list of
preloads for the organization.  However, to bring it in line with the
other functions of the module, just have it accept a list of `opts`
which can contain `preloads` but is a Keyword list.