Add Players.Ratings.scale_attributes/2
Add a method that will allow the caller to change the scale of the
player ratings, from their default scale of 1-200 to whichever scale
they desire.  However, the result of the scaling is always rounded up
to the next nearest integer.

I understand that the rounding up sounds strange, but this is actually a
feature of the game, trying to create some "fogginess" about the
player's abilities compared to one another.

For example, lets say that your league has decided to use the 1-10
ratings scale, and you have two hitters, whose `home_run_power` ratings
are 141 and 159 on the 200 scale.  While that is in reality more than a
12% advantage, scale to an integer on the 1-10 scale, both players
evalulate as a 8:

141 * 10 (NEW_SCALE) / 200 (OLD_SCALE) = 7.05 then round up to 8
159 * 10 (NEW_SCALE) / 200 (OLD_SCALE) = 7.95 then round up to 8