Refactor the way Player attributes are modeled
Rename the concept of Players.Ratings to Players.Attributes, and instead
of deliniating them by the batting/pitching/fielding/position tables,
just create one table called `player_attributes` which stores the name
of the attribute along with its type.  Note that on the _import_ side,
we still break them down into their sub types, because that is the way
they are stored on disk, but unite them in our application database.

Then add methods to `Players.Attributes`, to dilenate which attributes
belong to which `category` so that the Player model can have
associations for each "type" of attributes.