Player.Attirbutes component shows more attributes
Update the Player.Attributes component to display more than just the
Primary attributes, now show a Attributes.Secondary component, which
renders a unique set depending if the player being rendered is a hitter
or a pitcher.

This is a _very_ rough implementation, in relaity, I'd rather see the
Attributes component just render a single list of attribute comopnents
that is just `flex-wrapped`, but this is a start.