Order `@pitches` as reference for displaying
When dealing with pitches, OOTP has a standard order that they always
display them in, and I think it would be disorienting to not dipslay
them with this same logic.

So, reorder the `@pitches` attribute and use it as the source of truth
for the order that pitches attributes should be displayed in.