Setup Imports as a supervised OTP process tree
Mostly to gain more experience with designing OTP applications, setup
the import process as a OTP application.  This involves creating an
ImportSupervisor which oversees the following:

Import Agent - Stores the IDs of the parent records (Games, Teams,
Players, etc...) which are needed to validate dependent records when
they are imported.

Import Worker Pool - A pool of ImportWorkers, which are used to perform
the actual import, but to prevent trying to open too many DB connections
simultaneously they are capped at a fixed number of workers.

ImportState - A GenServer responsible for tracking the current state of
an import, one per session

ImportSession - A GenServer that represents an instance of an import,
which tracks the graph of import dependents and has is own state.